# docs/openapi.yaml
openapi: 3.1.0
info:
  title: kb-partners API
  version: 0.1.0
  description: >-
    Модуль обработки партнёрских заявок reHome (ТЗ v1.1). Contract-first:
    эта спецификация — источник истины для контрактных тестов (AT-002).
    На M0 (каркас) определены только инфраструктурные эндпоинты; доменные
    пути (/requests, /channels, /inbound, /webhooks) добавляются по эпикам
    M1–M7 под префиксом /api/v1/partners.
  contact:
    name: ООО РЕХОМ
servers:
  - url: https://api.rehome.one
    description: Production
  - url: https://staging-api.rehome.one
    description: Staging
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/partners/channels:
    get:
      tags: [Channels]
      summary: Список конфигураций каналов
      operationId: listChannels
      parameters:
        - name: collaborator_id
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Список каналов.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChannelConfigRead"
        "403":
          description: Требуется admin-скоуп.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags: [Channels]
      summary: Создать конфигурацию канала
      operationId: createChannel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelConfigCreate"
      responses:
        "201":
          description: Канал создан.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelConfigRead"
        "403":
          description: Требуется admin-скоуп.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Канал для этого партнёра и типа уже существует.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/partners/channels/{config_id}:
    parameters:
      - name: config_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Channels]
      summary: Карточка конфигурации канала
      operationId: getChannel
      responses:
        "200":
          description: Конфигурация канала.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelConfigRead"
        "404":
          description: Не найдено.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      tags: [Channels]
      summary: Изменить конфигурацию канала
      operationId: updateChannel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelConfigUpdate"
      responses:
        "200":
          description: Обновлённая конфигурация.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelConfigRead"
        "404":
          description: Не найдено.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak RS256 Bearer JWT (aud содержит `kb-partners`).
  schemas:
...
    Error:
      description: RFC 7807 problem+json (без ПДн в detail).
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        errors:
          type: array
          items:
            type: object
            additionalProperties:
              type: string

    # --- Web-push подписки (E8, FR-10.1) ---
...
    ChannelType:
      type: string
      enum: [API, CRM, TELEGRAM, MAX, EMAIL, MOCK]
...
    ChannelConfigCreate:
      description: Конфигурация канала партнёра. inbound_token — секрет, в выдаче не возвращается.
      type: object
      additionalProperties: false
      required: [collaborator_id, channel_type]
      properties:
        collaborator_id:
          type: string
          minLength: 1
          maxLength: 255
        channel_type:
          $ref: "#/components/schemas/ChannelType"
        priority:
          type: integer
          minimum: 0
          default: 100
        config:
          type: object
          additionalProperties: true
        inbound_token:
          type: string
          maxLength: 255
        is_active:
          type: boolean
          default: true
    ChannelConfigUpdate:
      description: Частичное обновление канала.
      type: object
      additionalProperties: false
      properties:
        priority:
          type: integer
          minimum: 0
        config:
          type: object
          additionalProperties: true
        inbound_token:
          type: string
          maxLength: 255
        is_active:
          type: boolean
    ChannelConfigRead:
      description: Представление канала. БЕЗ inbound_token (секрет).
      type: object
      additionalProperties: false
      required:
        [id, collaborator_id, channel_type, priority, config, is_active, health, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        collaborator_id:
          type: string
        channel_type:
          $ref: "#/components/schemas/ChannelType"
        priority:
          type: integer
        config:
          type: object
          additionalProperties: true
        is_active:
          type: boolean
        health:
          oneOf:
            - type: object
              additionalProperties: true
            - type: "null"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time